refactor: extract ExternalPrefabStageChangeTracker from ExternalSceneChangeTracker
Extract Class: move all Prefab Stage snapshot tracking (event
registration, save/reload, focus-return resolution, reopen-context
decision) into its own tracker, mirroring the existing Scene tracking
responsibility split. ExternalSceneChangeTracker keeps Scene tracking
and delegates Prefab Stage concerns to the new class; shared
fingerprint/path/logging helpers are now internal so both classes can
share them without behavior change.
