+++ Packages/src/Editor/FirstPartyTools/Compile/ExternalPrefabStageChangeTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40a74b393611486384404ee0c20e00d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
